+++ src/SumatraPDF.cpp
@@ -10990,6 +10990,10 @@ static LRESULT FrameOnCommand(MainWindow* win, HWND hwnd, UINT msg, WPARAM wp, L
             ShowConvertToPdfDialog(win);
             break;
 
+        case CmdConvertPdfToImages:
+            ShowConvertPdfToImagesDialog(win);
+            break;
+
         case CmdPdfCompress:
             ShowPdfCompressDialog(win);
             break;